requirements.txt
torch>=2.0.0
numpy>=1.21.0
python-chess>=1.0.0
tqdm>=4.65.0
pygame>=2.1.0
pytest>=7.0.0
pytest-cov>=4.0.0
